+++ src/finding/mod.rs
@@ -2,11 +2,14 @@
 // Licensed under the MIT License
 // SPDX-License-Identifier: MIT
 
+pub mod call_site_finder;
 pub mod implemented_type;
 pub mod implemented_type_finder;
 pub mod import_path;
 pub mod module_declaration_finder;
 pub mod package_tree;
+pub mod public_entry_point;
+pub mod public_entry_point_finder;
 pub mod qualified_call;
 pub mod qualified_call_finder;
 pub mod registry_item;