+++ tests/finding/mod.rs
@@ -2,11 +2,14 @@
 // Licensed under the MIT License
 // SPDX-License-Identifier: MIT
 
+pub mod call_site_finder_tests;
 pub mod implemented_type_finder_tests;
 pub mod implemented_type_tests;
 pub mod import_path_tests;
 pub mod module_declaration_finder_tests;
 pub mod package_tree_tests;
+pub mod public_entry_point_finder_tests;
+pub mod public_entry_point_tests;
 pub mod qualified_call_finder_tests;
 pub mod qualified_call_tests;
 pub mod registry_item_tests;